requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
comm==0.1.3
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.8
distro==1.8.0
executing==1.2.0
filelock==3.13.1
frozenlist==1.4.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.1
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
opencv-python==4.9.0.80
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.14.0
pyparsing==3.1.1
python-dateutil==2.8.2
pyzmq==25.0.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.25.1
wcwidth==0.2.6
yarl==1.9.2
